Shelter Navigation fixed
